refactor(axum-http-server): split test contract module into directory module
Split the large contract.rs (2018 lines) into smaller per-concern modules
to improve maintainability and make it easier for AI agents to work with.

- for_all_config_modes: further split into and_running_on_reverse_proxy,
  receiving_an_announce_request, receiving_an_scrape_request
- configured_as_whitelisted
- configured_as_private
- configured_as_private_and_whitelisted (with TODOs for missing tests)
- using_ipv6_v6only
