+++ packages/axum-http-server/tests/server/v1/contract/configured_as_private_and_whitelisted.rs
@@ -1,0 +1,9 @@
+mod and_receiving_an_announce_request {
+    // TODO: add tests for announce requests when the tracker is configured as both private and whitelisted.
+    //       See `configured_as_private` and `configured_as_whitelisted` modules for the individual test patterns.
+}
+
+mod receiving_an_scrape_request {
+    // TODO: add tests for scrape requests when the tracker is configured as both private and whitelisted.
+    //       See `configured_as_private` and `configured_as_whitelisted` modules for the individual test patterns.
+}
